♻️ refactor: make `coordinaxs.api` the canonical home for `CKey`/`CDict`
`coordinaxs.api` is the root of the workspace -- every other package depends on
it, directly or through `coordinax`, and it depends on none of them. That makes
it the one place all layers can import from, so it is where the shared
component-key vocabulary belongs.

Rename `coordinaxs/api/_custom_types.py` to `custom_types.py` and make it the
sole definition; `coordinax._src.custom_types` re-exports. The package has no
`__init__.py` by design, so the module path is the public path:

    from coordinaxs.api.custom_types import CDict, CKey

This removes the last duplicate rather than documenting it. `coordinaxs.api`
could never re-export *from* `coordinax`, but it can be the source -- the arrow
still points `coordinaxs.api <- coordinax`.

`coordinax.internal` no longer exports `CDict`/`CKey`; consumers use the public
path above. `Shape`/`HasShape` are unaffected and stay canonical in
`coordinax._src.custom_types`.

`coordinaxs.hypothesis` gains the `coordinaxs.api` dependency it now imports
directly, instead of relying on it transitively through `coordinax`.
