+++ packages/coordinaxs.interop.astropy/src/coordinaxs/interop/astropy/_src/custom_types.py
@@ -2,4 +2,4 @@
 
 __all__ = ("CDict",)
 
-from coordinax.internal import CDict
+from coordinaxs.api.custom_types import CDict
