+++ src/coordinax/vectors/_src/register_cx.py
@@ -22,7 +22,7 @@ import coordinaxs.api.transforms as cxfmapi
 from .bundle import Coordinate
 from .point import Point
 from .tangent import Tangent
-from coordinax.internal import CDict, OptUSys
+from coordinax._src.custom_types import CDict, OptUSys
 
 CHART_MSMTCH = "from_chart {0} does not match the point's chart {1.chart}"
