+++ src/coordinax/vectors/_src/tangent.py
@@ -23,7 +23,7 @@ from .mixins import AstropyRepresentationAPIMixin
 from .point import _frame_converter
 
 if TYPE_CHECKING:
-    from coordinax.internal import CDict
+    from coordinax._src.custom_types import CDict
 
 ChartT = TypeVar(
     "ChartT",
